Default build options to generate table shell

Default build options to generate table shell

defaultTableShellBuildOptions(
  codesetTempTable = "#codeset",
  sourceCodesetTempTable = "#source_codeset",
  timeWindowTempTable = "#time_windows",
  targetCohortTempTable = "#target_cohorts",
  tsMetaTempTable = "#ts_meta",
  conceptSetOccurrenceTempTable = "#concept_set_occ",
  cohortOccurrenceTempTable = "#cohort_occ",
  patientLevelDataTempTable = "#patient_data",
  patientLevelTableShellTempTable = "#pat_ts_tab",
  categoricalSummaryTempTable = "#categorical_table",
  continuousSummaryTempTable = "#continuous_table",
  cohortAnalysisType = c("era", "startDate")
)

Arguments

codesetTempTable

the name of the codeset table used in execution. Defaults as a temp table #codeset

timeWindowTempTable

the name of the time Window table used in execution. Defaults as a temp table #time_windows

targetCohortTempTable

the name of the target cohort table used in execution. Defaults as a temp table #target_cohorts

tsMetaTempTable

the name of the table shell meta table used in execution. Defaults as a temp table #ts_meta

conceptSetOccurrenceTempTable

the name of the concept set occurrence table used in execution. Defaults as a temp table #concept_set_occ

cohortOccurrenceTempTable

the name of the cohort occurrence table used in execution. Defaults as a temp table #cohort_occ

patientLevelDataTempTable

the name of the patient level data table used in execution. Note this does not contain info of the table shell. Defaults as a temp table #patient_data

patientLevelTableShellTempTable

the name of the patient level data table with additional meta info used in execution. Defaults as a temp table #pat_ts_tab

categoricalSummaryTempTable

the name of the categorical summary table used in execution. Defaults as a temp table #categorical_table

continuousSummaryTempTable

the name of the continuous summary table used in execution. Defaults as a temp table #continuous_table

connectionDetails

A DatabaseConnector connectionDetails object (optional if connection is specified)

useCohortEra

a true false toggle specifying if in a cohort Char whether to use the cohort era (TRUE) or just the start date (FALSE)

Value

A BuildOptions object